Creating dedicated file for CPU unit tests
